Added Updated Project Folder
Because of our use of external open source applications in building out
own pedometer implementation, the final implementation along with an
accompanying GUI have been added to the main folder as a single project
folder. The contents of this folder will ultimately make up the entirety
of the finished project.